requirements.txt
bs4
requests
lxml
selenium
webdriver_manager
fake-useragent
aiogram
python-decouple
